# feat: Implement comprehensive authentication system with JWT, email verification, password reset, and 90% test coverage