--- a/src/controllers/AuthController.ts
+++ b/src/controllers/AuthController.ts
@@ -17,7 +17,21 @@ export class AuthController {
   }
 
   /**
-   * Register a new user
+   * Get client IP address from request
+   */
+  private getClientIp(req: Request): string {
+    return (
+      (req.headers['x-forwarded-for'] as string) ||
+      (req.headers['x-real-ip'] as string) ||
+      req.connection.remoteAddress ||
+      req.socket.remoteAddress ||
+      req.ip ||
+      'unknown'
+    );
+  }
+
+  /**
+   * Register a new user with enhanced security
    */
   register = asyncHandler(async (req: Request, res: Response): Promise<void> => {
     logger.info('Registration request received');
@@ -32,12 +46,15 @@ export class AuthController {
       return;
     }
 
-    // Register user
+    // Register user with email verification
     const newUser = await this.authService.register(sanitizedData);
 
     // Format response
     const userData = AuthResource.formatRegisterResponse(newUser);
-    const successResponse = createSuccessResponse('User registered successfully', userData);
+    const successResponse = createSuccessResponse(
+      'User registered successfully. Please check your email for verification.',
+      userData
+    );
 
     logger.info(`User ${newUser.id} registered successfully`);
 
@@ -45,7 +62,7 @@ export class AuthController {
   });
 
   /**
-   * Login user
+   * Login user with enhanced security
    */
   login = asyncHandler(async (req: Request, res: Response): Promise<void> => {
     logger.info('Login request received');
@@ -60,12 +77,18 @@ export class AuthController {
       return;
     }
 
-    // Login user
-    const loginData = await this.authService.login({
-      email: sanitizedData.email,
-      password: sanitizedData.password,
-      ...(sanitizedData.remember_me !== undefined && { rememberMe: sanitizedData.remember_me }),
-    });
+    // Get client IP for rate limiting
+    const clientIp = this.getClientIp(req);
+
+    // Login user with enhanced security
+    const loginData = await this.authService.login(
+      {
+        email: sanitizedData.email,
+        password: sanitizedData.password,
+        ...(sanitizedData.remember_me !== undefined && { rememberMe: sanitizedData.remember_me }),
+      },
+      clientIp
+    );
 
     // Get full user data for response
     const user = await this.authService.getUserFromToken(loginData.tokens.accessToken);
@@ -90,6 +113,58 @@ export class AuthController {
     const successResponse = createSuccessResponse('Login successful', responseData);
 
     logger.info(`User ${user.id} logged in successfully`);
+
+    res.status(200).json(successResponse);
+  });
+
+  /**
+   * Logout user
+   */
+  logout = asyncHandler(async (req: Request, res: Response): Promise<void> => {
+    logger.info('Logout request received');
+
+    // Get user ID from JWT token
+    const token = req.headers.authorization?.replace('Bearer ', '');
+    if (!token) {
+      res.status(401).json({ error: 'No token provided' });
+      return;
+    }
+
+    const user = await this.authService.getUserFromToken(token);
+    if (!user) {
+      res.status(401).json({ error: 'Invalid token' });
+      return;
+    }
+
+    // Logout user
+    await this.authService.logout(user.id);
+
+    const successResponse = createSuccessResponse('Logout successful');
+
+    logger.info(`User ${user.id} logged out successfully`);
+
+    res.status(200).json(successResponse);
+  });
+
+  /**
+   * Refresh JWT tokens
+   */
+  refreshTokens = asyncHandler(async (req: Request, res: Response): Promise<void> => {
+    logger.info('Token refresh request received');
+
+    const { refresh_token } = req.body;
+
+    if (!refresh_token) {
+      res.status(400).json({ error: 'Refresh token is required' });
+      return;
+    }
+
+    // Refresh tokens
+    const refreshData = await this.authService.refreshTokens(refresh_token);
+
+    const successResponse = createSuccessResponse('Tokens refreshed successfully', refreshData);
+
+    logger.info('Tokens refreshed successfully');
 
     res.status(200).json(successResponse);
   });