--- a/src/database/migrations/20250805083029_user_sessions.ts
+++ b/src/database/migrations/20250805083029_user_sessions.ts
@@ -1,0 +1,25 @@
+import type { Knex } from 'knex';
+
+export async function up(knex: Knex): Promise<void> {
+  return knex.schema.createTable('user_sessions', table => {
+    table.uuid('id').primary().defaultTo(knex.raw('gen_random_uuid()'));
+    table.uuid('user_id').notNullable().references('id').inTable('users').onDelete('CASCADE');
+    table.string('refresh_token', 500).notNullable().unique();
+    table.string('device_info', 500);
+    table.specificType('ip_address', 'INET');
+    table.timestamp('expires_at').notNullable();
+    table.boolean('is_active').defaultTo(true);
+    table.timestamp('last_used_at').defaultTo(knex.fn.now());
+    table.timestamp('created_at').defaultTo(knex.fn.now());
+    table.timestamp('updated_at').defaultTo(knex.fn.now());
+
+    // Indexes
+    table.index(['user_id']);
+    table.index(['refresh_token']);
+    table.index(['expires_at']);
+  });
+}
+
+export async function down(knex: Knex): Promise<void> {
+  return knex.schema.dropTable('user_sessions');
+}