--- a/src/database/migrations/20250805120000_add_email_verified_to_users.ts
+++ b/src/database/migrations/20250805120000_add_email_verified_to_users.ts
@@ -1,0 +1,13 @@
+import type { Knex } from 'knex';
+
+export async function up(knex: Knex): Promise<void> {
+  return knex.schema.alterTable('users', table => {
+    table.boolean('email_verified').defaultTo(false).notNullable();
+  });
+}
+
+export async function down(knex: Knex): Promise<void> {
+  return knex.schema.alterTable('users', table => {
+    table.dropColumn('email_verified');
+  });
+}