--- a/src/database/migrations/20250805120004_alter_verification_token_column.ts
+++ b/src/database/migrations/20250805120004_alter_verification_token_column.ts
@@ -1,0 +1,15 @@
+import type { Knex } from 'knex';
+
+export async function up(knex: Knex): Promise<void> {
+  return knex.schema.alterTable('user_email_verifications', table => {
+    // Change verification_token from VARCHAR(255) to TEXT to accommodate JWT tokens
+    table.text('verification_token').alter();
+  });
+}
+
+export async function down(knex: Knex): Promise<void> {
+  return knex.schema.alterTable('user_email_verifications', table => {
+    // Revert back to VARCHAR(255)
+    table.string('verification_token', 255).alter();
+  });
+}