… package.json
… src/controllers/PostController.ts
… src/mails/passwordResetEmail.ts
… src/mails/verificationEmail.ts
… src/mails/welcomeEmail.ts
… src/middleware/rateLimiter.ts
… src/models/BaseModel.ts
… src/models/PostModel.ts
… src/models/UserModel.ts
… src/models/index.ts
… src/resources/authResource.ts
… src/resources/loginResource.ts
… src/routes/auth.ts
… src/routes/users.ts
… src/services/AuditLogService.ts
… src/services/AuthService.ts
… src/services/EmailVerificationService.ts
… src/services/LoginAttemptService.ts
… src/services/PostService.ts
… src/services/UserService.ts
… src/tests/README.md
… src/tests/integration/auth-endpoints.test.ts
… src/tests/performance/performance.test.ts
… src/tests/security/security.test.ts
… src/tests/services/AuthService.test.ts
… src/tests/services/UserService.test.ts
… src/tests/setup.ts
… src/tests/utils/jwt.test.ts
… src/tests/validators/profileValidator.test.ts
… src/types/auth.ts